+++ numerapi/numerapi.py
@@ -937,8 +937,12 @@ class NumerAPI(base_api.Api):
         submission_resp = self.raw_query(auth_query, arguments,
                                          authorization=True)
         submission_auth = submission_resp['data']['submission_upload_auth']
+
+        # get compute id if available and pass it along
+        headers = {"x_compute_id": os.getenv("NUMERAI_COMPUTE_ID")}
         with open(file_path, 'rb') as fh:
-            requests.put(submission_auth['url'], data=fh.read())
+            requests.put(
+                submission_auth['url'], data=fh.read(), headers=headers)
         create_query = '''
             mutation($filename: String!
                      $tournament: Int!